google_ads_wrapper/google-ads.py
import functions_framework
import json
import pandas as pd
from google.ads.googleads.client import GoogleAdsClient
from google.oauth2.credentials import Credentials
from pandas_gbq import to_gbq
import requests

@functions_framework.http
def google_ads_data_function(request):
    """
    Cloud Function to handle Google Ads data fetching and uploading to BigQuery.

    Args:
        request: The HTTP request containing JSON payload with required parameters.

    Returns:
        Response message indicating success or failure of the operation.
    """
    request_json = request.get_json(silent=True)
    if not request_json:
        return "Invalid request. Please send a valid JSON payload.", 400

    # Parse required parameters
    try:
        start_date = request_json["start_date"]
        end_date = request_json["end_date"]
        refresh_token = request_json["refresh_token"]
        report_name = request_json["report_name"]
        login_customer_id = request_json["login_customer_id"]
        job_id = request_json.get("job_id")
    except KeyError:
        return "Missing required parameters in the request.", 400

    # Update job status to inProgress
    update_job_status(job_id, "inProgress")

    try:
        # Create Google Ads client
        client = create_client(refresh_token, login_customer_id)
        client_accounts = list_client_accounts(client, login_customer_id)

        if not client_accounts:
            update_job_status(job_id, "failed")
            return "No client accounts found.", 404

        account_id = client_accounts[0]
        report_data = get_google_ads_data(client, account_id, start_date, end_date, report_name)

        # Send data to BigQuery
        dataset_name = "your_dataset_name"  # Replace with actual dataset name
        project_id = "your_project_id"      # Replace with your GCP project ID
        send_to_bigquery(report_data, report_name, dataset_name, project_id)

        update_job_status(job_id, "success")
        return json.dumps({
            "message": f"Data for {report_name} successfully fetched and uploaded to BigQuery",
            "rows_uploaded": len(report_data)
        }), 200
    except Exception as e:
        update_job_status(job_id, "failed")
        return json.dumps({"error": str(e)}), 500


def update_job_status(job_id, status):
    """Update the job status via an API."""
    try:
        url = "https://your-job-status-api.com/update"  # Replace with actual URL
        payload = {"jobId": job_id, "status": status}
        headers = {"Content-Type": "application/json"}
        requests.post(url, json=payload, headers=headers)
    except Exception as e:
        print(f"Failed to update job status: {e}")


def create_client(refresh_token, login_customer_id):
    """Create a Google Ads API client."""
    developer_token = "your_developer_token"  # Replace with actual developer token
    client_id = "your_client_id"              # Replace with your client ID
    client_secret = "your_client_secret"      # Replace with your client secret

    credentials = Credentials.from_authorized_user_info(
        {
            "client_id": client_id,
            "client_secret": client_secret,
            "refresh_token": refresh_token,
        }
    )
    client = GoogleAdsClient(credentials=credentials, developer_token=developer_token)
    client.login_customer_id = login_customer_id
    return client


def list_client_accounts(client, manager_customer_id):
    """Fetch client accounts under the manager account."""
    # Implementation here
    pass


def get_google_ads_data(client, account_id, start_date, end_date, report_type):
    """Fetch Google Ads data based on report type."""
    # Implementation here
    pass


def send_to_bigquery(df, report_name, dataset_name, project_id):
    """Send the DataFrame to BigQuery."""
    if not df.empty:
        full_table_name = f"{dataset_name}.{report_name}"
        to_gbq(df, full_table_name, project_id=project_id, if_exists="append")
