google_ads_wrapper/requirements.txt
functions-framework
google-ads
pandas
pandas-gbq
google-cloud-bigquery
requests
google-auth
google-auth-oauthlib